Intcode computer uses longs

+++ 2019/src/main/java/advent2019/Day2.java
@@ -16,7 +16,7 @@ public final class Day2 {
         System.out.println(part2(Files.lines(INPUT_PATH)));
     }
 
-    private static int part1(Stream<String> lines) {
+    private static long part1(Stream<String> lines) {
         IntcodeComputer computer = IntcodeComputer.builder()
                 .initialMemory(lines.findFirst().orElseThrow())
                 .build();
